moving a module to a submodule.
